+++ CHANGELOG.md
@@ -4,10 +4,11 @@ Note the first digit of every adapter version corresponds to the major version o
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
 ### 4.7.0.0.1
+- Add support for Adaptive Banners.
 - This version of the adapters has been certified with VungleAds 7.0.0.
 
 ### 4.7.0.0.0
-- This version of the adapters has been certified with VungleSDK-iOS 7.0.0.
+- This version of the adapters has been certified with VungleSDK 7.0.0.
 
 ### 4.6.12.3.1
 - Added support for the Rewarded Interstitial ad format.